requirements.txt
wget
ffmpeg-python
fastapi
uvicorn